requirements.txt
openenv-core>=0.2.2
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
openai>=1.30.0
huggingface-hub>=0.23.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
